+++ package.json
@@ -74,28 +74,25 @@
     "git-rev-sync": "^1.12.0",
     "husky": "^0.15.0-rc.8",
     "lint-staged": "^7.0.5",
-    "madge": "^3.1.1",
+    "madge": "^3.4.4",
     "md5-file": "^4.0.0",
     "mkdirp": "^0.5.1",
     "mocha": "^5.1.1",
     "multer": "^1.3.0",
     "node-fetch": "^2.3.0",
-    "nyc": "^11.7.1",
-    "parse5": "^4.0.0",
-    "prettier": "^1.14.0",
+    "nyc": "^13.3.0",
+    "prettier": "^1.16.4",
     "query-string": "^5.0.1",
-    "simple-peer": "^8.5.0",
     "sinon": "^6.1.3",
     "sinon-chai": "^3.3.0",
     "titere": "^2.0.3",
-    "tslint": "^5.11.0",
+    "tslint": "^5.14.0",
     "tslint-language-service": "^0.9.9",
     "typescript": "^3.2.2",
     "uuid": "^3.1.0"
   },
   "dependencies": {
     "@createjs/easeljs": "^2.0.0-beta.4",
-    "@types/simple-peer": "^6.1.4",
     "babylonjs": "^3.3.0",
     "babylonjs-gltf2interface": "^3.3.0",
     "babylonjs-gui": "^3.3.0",
@@ -104,20 +101,19 @@
     "babylonjs-materials": "^3.3.0",
     "babylonjs-procedural-textures": "^3.3.0",
     "body-scroll-lock": "^2.5.10",
-    "cids": "^0.5.5",
-    "concurrently": "^4.0.1",
+    "concurrently": "^4.1.0",
     "decentraland-rpc": "^3.1.5",
     "devtools-protocol": "0.0.615714",
     "ephemeralkey": "^2.0.1",
-    "eth-connect": "^4.0.0",
+    "eth-connect": "^4.0.1",
     "ethereum-blockies": "^0.1.1",
     "fp-future": "^1.0.1",
     "google-protobuf": "^3.6.1",
     "hammerjs": "^2.0.8",
-    "keccakjs": "^0.2.1",
+    "keccakjs": "^0.2.3",
     "multihashes": "^0.4.14",
     "oimo": "^1.0.9",
-    "uglify-js": "^3.4.9",
+    "uglify-js": "^3.5.2",
     "webrtc-adapter": "^6.4.8"
   }
 }
