--- packages/engine/entities/utils/animationSystem.ts
@@ -1,9 +1,0 @@
-import * as BABYLON from 'babylonjs'
-import { BaseEntity } from '../BaseEntity'
-
-/// --- EXPORTS ---
-
-export interface IAnimatedEntity extends BaseEntity {
-  skeletons: BABYLON.Skeleton[]
-  animationGroups: BABYLON.AnimationGroup[]
-}
