+++ packages/engine/renderer/camera.ts
@@ -2,19 +2,12 @@ import * as BABYLON from 'babylonjs'
 
 import { vrHelper, scene, engine } from './init'
 import { playerConfigurations, visualConfigurations, parcelLimits } from 'config'
-import { keyState, Keys } from './input'
 import { gridToWorld } from 'atomicHelpers/parcelScenePositions'
 import { teleportObservable } from 'shared/world/positionThings'
 
+const vrCamera = vrHelper.deviceOrientationCamera as BABYLON.DeviceOrientationCamera
 export const DEFAULT_CAMERA_ZOOM = 32
 export const MAX_CAMERA_ZOOM = 100
-const CAMERA_SPEED = 0.02
-const CAMERA_LEFT = BABYLON.Quaternion.RotationYawPitchRoll(Math.PI / 2, 0, 0)
-const CAMERA_RIGHT = BABYLON.Quaternion.RotationYawPitchRoll(-Math.PI / 2, 0, 0)
-const CAMERA_FORWARD = BABYLON.Quaternion.RotationYawPitchRoll(Math.PI, 0, 0)
-const CAMERA_BACKWARD = BABYLON.Quaternion.RotationYawPitchRoll(0, 0, 0)
-
-const vrCamera = vrHelper.deviceOrientationCamera as BABYLON.DeviceOrientationCamera
 
 const arcCamera = new BABYLON.ArcRotateCamera(
   'arc-camera',
@@ -46,31 +39,6 @@ const arcCamera = new BABYLON.ArcRotateCamera(
 
   vrHelper.position = vrCamera.position
 
-  arcCamera.keysDown = []
-  arcCamera.keysUp = []
-  arcCamera.keysLeft = []
-  arcCamera.keysRight = []
-
-  arcCamera.onAfterCheckInputsObservable.add(() => {
-    if (arcCamera === scene.activeCamera) {
-      if (keyState[Keys.KEY_LEFT]) {
-        arcCamera.target.addInPlace(moveCamera(arcCamera, CAMERA_LEFT, CAMERA_SPEED * engine.getDeltaTime()))
-      }
-
-      if (keyState[Keys.KEY_RIGHT]) {
-        arcCamera.target.addInPlace(moveCamera(arcCamera, CAMERA_RIGHT, CAMERA_SPEED * engine.getDeltaTime()))
-      }
-
-      if (keyState[Keys.KEY_UP]) {
-        arcCamera.target.addInPlace(moveCamera(arcCamera, CAMERA_FORWARD, CAMERA_SPEED * engine.getDeltaTime()))
-      }
-
-      if (keyState[Keys.KEY_DOWN]) {
-        arcCamera.target.addInPlace(moveCamera(arcCamera, CAMERA_BACKWARD, CAMERA_SPEED * engine.getDeltaTime()))
-      }
-    }
-  })
-
   arcCamera.upperBetaLimit = Math.PI / 2
   arcCamera.allowUpsideDown = false
   arcCamera.upperRadiusLimit = arcCamera.panningDistanceLimit = MAX_CAMERA_ZOOM
@@ -82,40 +50,6 @@ const arcCamera = new BABYLON.ArcRotateCamera(
 }
 
 /// --- EXPORTS ---
-
-function applyQuaternion(v: BABYLON.Vector3, q: BABYLON.Quaternion) {
-  let x = v.x
-  let y = v.y
-  let z = v.z
-  let qx = q.x
-  let qy = q.y
-  let qz = q.z
-  let qw = q.w
-
-  // calculate quat * vector
-
-  let ix = qw * x + qy * z - qz * y
-  let iy = qw * y + qz * x - qx * z
-  let iz = qw * z + qx * y - qy * x
-  let iw = -qx * x - qy * y - qz * z
-
-  // calculate result * inverse quat
-
-  v.x = ix * qw + iw * -qx + iy * -qz - iz * -qy
-  v.y = iy * qw + iw * -qy + iz * -qx - ix * -qz
-  v.z = iz * qw + iw * -qz + ix * -qy - iy * -qx
-
-  return v
-}
-
-function moveCamera(camera: BABYLON.ArcRotateCamera, directionRotation: BABYLON.Quaternion, speed: number) {
-  const direction = camera.position.subtract(camera.target)
-  direction.y = 0
-  direction.normalize()
-
-  applyQuaternion(direction, directionRotation)
-  return direction.scaleInPlace(speed)
-}
 
 export { vrCamera, arcCamera }
